+++ ydb/core/tx/columnshard/engines/reader/simple_reader/duplicates/interval_index.cpp
@@ -1,0 +1,4 @@
+#include "interval_index.h"
+
+namespace NKikimr::NOlap::NReader::NSimple {
+}
