+++ ydb/core/tx/columnshard/engines/reader/simple_reader/duplicates/manager.h
@@ -63,293 +63,6 @@ public:
     }
 };
 
-class TIntervalCounter {
-private:
-    class TModification {
-    private:
-        YDB_READONLY_DEF(ui32, Left);
-        YDB_READONLY_DEF(ui32, Right);
-        YDB_READONLY_DEF(i64, Delta);
-
-    public:
-        TModification(const ui32 l, const ui32 r, const i64 delta)
-            : Left(l)
-            , Right(r)
-            , Delta(delta) {
-        }
-    };
-
-    class TPosition {
-    private:
-        YDB_READONLY_DEF(ui32, Index);
-        YDB_READONLY_DEF(ui32, Left);
-        YDB_READONLY_DEF(ui32, Right);
-
-        TPosition(const ui32 i, const ui32 l, const ui32 r)
-            : Index(i)
-            , Left(l)
-            , Right(r) {
-        }
-
-    public:
-        explicit TPosition(const ui32 maxIndex)
-            : Index(0)
-            , Left(0)
-            , Right(maxIndex) {
-        }
-
-        TPosition LeftChild() const {
-            AFL_VERIFY(Right > Left);
-            return TPosition(Index * 2 + 1, Left, (Left + Right - 1) / 2);
-        }
-
-        TPosition RightChild() const {
-            AFL_VERIFY(Right > Left);
-            return TPosition(Index * 2 + 2, (Left + Right + 1) / 2, Right);
-        }
-
-        ui32 IntervalSize() const {
-            return Right - Left + 1;
-        }
-    };
-
-    class TZeroCollector {
-    private:
-        std::vector<ui32> Positions;
-
-    public:
-        void OnNewZeros(const TPosition& node) {
-            for (ui32 i = node.GetLeft(); i <= node.GetRight(); ++i) {
-                Positions.emplace_back(i);
-            }
-        }
-
-        std::vector<ui32> ExtractNewZeroPositions() {
-            return std::move(Positions);
-        }
-    };
-
-    // Segment tree: Count[i] = GetCount(i * 2 + 1) + GetCount(i * 2 + 2)
-    std::vector<ui64> Count;
-    std::vector<ui64> MinValue;
-    std::vector<i64> PropagatedDeltas;
-    ui32 MaxIndex = 0;
-
-private:
-    void PropagateDelta(const TPosition& node);
-    void Update(const TPosition& node, const TModification& modification, TZeroCollector* callback);
-    void Inc(const ui32 l, const ui32 r);
-    ui64 GetCount(const TPosition& node, const ui32 l, const ui32 r);
-    ui64 GetCount(const TPosition& node) const {
-        AFL_VERIFY(node.GetIndex() < Count.size());
-        return Count[node.GetIndex()] +
-               (node.GetIndex() < PropagatedDeltas.size() ? PropagatedDeltas[node.GetIndex()] : 0) * node.IntervalSize();
-    }
-    ui64 GetMinValue(const TPosition& node) const {
-        AFL_VERIFY(node.GetIndex() < MinValue.size());
-        return MinValue[node.GetIndex()] + (node.GetIndex() < PropagatedDeltas.size() ? PropagatedDeltas[node.GetIndex()] : 0);
-    }
-    TPosition GetRoot() const {
-        return TPosition(MaxIndex);
-    }
-    TString DebugString(const TPosition& node) const {
-        return TStringBuilder() << "node=" << node.GetIndex() << ";max_index=" << MaxIndex << ";count=" << GetCount(node)
-                                << ";min=" << GetMinValue(node);
-    }
-    void OnNodeUpdated(const TPosition& node, TZeroCollector* callback);
-    bool IsLeaf(const TPosition& node) const {
-        return node.GetIndex() >= MaxIndex;
-    }
-
-public:
-    TIntervalCounter(const std::vector<std::pair<ui32, ui32>>& intervals);
-    std::vector<ui32> DecAndGetZeros(const ui32 l, const ui32 r);
-    ui64 GetCount(const ui32 l, const ui32 r) {
-        return GetCount(GetRoot(), l, r);
-    }
-    bool IsAllZeros() const;
-};
-
-class TIntervalSet {
-private:
-    using TPosition = ui32;
-    using TIntervalId = ui64;
-
-private:
-    class TBorder {
-    private:
-        YDB_READONLY_DEF(bool, IsEnd);
-        YDB_READONLY_DEF(TPosition, Position);
-        YDB_READONLY_DEF(TIntervalId, Interval);
-
-        TBorder(const TPosition position, const TIntervalId interval, const bool isEnd)
-            : IsEnd(isEnd)
-            , Position(position)
-            , Interval(interval) {
-        }
-
-    public:
-        static TBorder Begin(const TPosition position, const TIntervalId interval) {
-            return TBorder(position, interval, false);
-        }
-        static TBorder End(const TPosition position, const TIntervalId interval) {
-            return TBorder(position, interval, true);
-        }
-
-        std::partial_ordering operator<=>(const TBorder& other) const {
-            return std::tie(IsEnd, Position, Interval) <=> std::tie(other.IsEnd, Position, other.Interval);
-        }
-    };
-
-public:
-    class TCursor;
-
-private:
-    using TIntervalsByPosition = std::map<TPosition, THashSet<TIntervalId>>;
-    std::set<TBorder> Borders;
-    std::set<std::weak_ptr<TCursor>, std::owner_less<std::weak_ptr<TCursor>>> Cursors;
-    TIntervalsByPosition IntervalsByLeft;
-    TIntervalsByPosition IntervalsByRight;
-
-public:
-    class TCursor {
-    private:
-        using TIdsByIdx = THashMap<ui32, std::vector<ui64>>;
-
-        const TIntervalSet* Container;
-        YDB_READONLY(ui32, IntervalIdx, 0);
-        YDB_READONLY_DEF(THashSet<ui64>, IntersectingSources);
-
-        void NextImpl(const bool init) {
-            if (!init) {
-                if (const auto findIntervals = Container->IntervalsByRight.find(IntervalIdx);
-                    findIntervals != Container->IntervalsByRight.end()) {
-                    for (const TIntervalId source : findIntervals->second) {
-                        AFL_VERIFY(IntersectingSources.erase(source));
-                    }
-                }
-
-                ++IntervalIdx;
-            }
-
-            if (const auto findIntervals = Container->IntervalsByLeft.find(IntervalIdx); findIntervals != Container->IntervalsByLeft.end()) {
-                for (const TIntervalId source : findIntervals->second) {
-                    AFL_VERIFY(IntersectingSources.emplace(source).second);
-                }
-            }
-        }
-
-        friend TIntervalSet;
-        TCursor(const TIntervalSet* container)
-            : Container(container) {
-            AFL_VERIFY(Container);
-            NextImpl(true);
-        }
-
-    public:
-        void Next() {
-            NextImpl(false);
-        }
-
-        void Prev() {
-            AFL_VERIFY(IntervalIdx);
-            if (const auto findIntervals = Container->IntervalsByLeft.find(IntervalIdx); findIntervals != Container->IntervalsByLeft.end()) {
-                for (const TIntervalId source : findIntervals->second) {
-                    AFL_VERIFY(IntersectingSources.erase(source));
-                }
-            }
-            --IntervalIdx;
-            if (const auto findIntervals = Container->IntervalsByRight.find(IntervalIdx); findIntervals != Container->IntervalsByRight.end()) {
-                for (const TIntervalId source : findIntervals->second) {
-                    AFL_VERIFY(IntersectingSources.emplace(source).second);
-                }
-            }
-        }
-
-        void Move(const ui64 intervalIdx) {
-            while (IntervalIdx < intervalIdx) {
-                Next();
-            }
-            while (IntervalIdx > intervalIdx) {
-                Prev();
-            }
-        }
-    };
-
-    template <typename F>
-    void UpdateCursors(const F& update) {
-        for (auto it = Cursors.begin(); it != Cursors.end();) {
-            if (auto locked = it->lock()) {
-                update(locked);
-                ++it;
-            } else {
-                it = Cursors.erase(it);
-            }
-        }
-    }
-
-public:
-    void Insert(const TPosition l, const TPosition r, const TIntervalId id) {
-        AFL_VERIFY(Borders.insert(TBorder::Begin(l, id)).second);
-        AFL_VERIFY(Borders.insert(TBorder::End(r, id)).second);
-
-        AFL_VERIFY(IntervalsByLeft[l].emplace(id).second);
-        AFL_VERIFY(IntervalsByRight[r].emplace(id).second);
-
-        const auto addInterval = [l, r, id](const std::shared_ptr<TCursor>& cursor) {
-            if (cursor->IntervalIdx >= l && cursor->IntervalIdx <= r) {
-                AFL_VERIFY(cursor->IntersectingSources.emplace(id).second);
-            }
-        };
-        UpdateCursors(addInterval);
-    }
-
-    void Erase(const TPosition l, const TPosition r, const TIntervalId id) {
-        AFL_VERIFY(Borders.erase(TBorder::Begin(l, id)));
-        AFL_VERIFY(Borders.erase(TBorder::End(r, id)));
-
-        {
-            auto findInterval = IntervalsByLeft.find(l);
-            AFL_VERIFY(findInterval != IntervalsByLeft.end());
-            AFL_VERIFY(findInterval->second.erase(id));
-            if (findInterval->second.empty()) {
-                IntervalsByLeft.erase(findInterval);
-            }
-        }
-        {
-            auto findInterval = IntervalsByRight.find(r);
-            AFL_VERIFY(findInterval != IntervalsByRight.end());
-            AFL_VERIFY(findInterval->second.erase(id));
-            if (findInterval->second.empty()) {
-                IntervalsByRight.erase(findInterval);
-            }
-        }
-
-        const auto eraseInterval = [l, r, id](const std::shared_ptr<TCursor>& cursor) {
-            if (cursor->IntervalIdx >= l && cursor->IntervalIdx <= r) {
-                AFL_VERIFY(cursor->IntersectingSources.erase(id));
-            }
-        };
-        UpdateCursors(eraseInterval);
-    }
-
-    THashSet<ui64> GetIntersections(const ui32 l, const ui32 r, const std::shared_ptr<TCursor>& cursor) const {
-        AFL_VERIFY(cursor->GetIntervalIdx() >= l && cursor->GetIntervalIdx() <= r);
-        THashSet<ui64> intervals = cursor->GetIntersectingSources();
-        for (auto it = Borders.lower_bound(TBorder::Begin(l, std::numeric_limits<TIntervalId>::min()));
-             it != Borders.end() && *it <= TBorder::End(r, std::numeric_limits<TIntervalId>::max()); ++it) {
-            intervals.insert(it->GetInterval());
-        }
-        return intervals;
-    }
-
-    std::shared_ptr<TCursor> MakeCursor() {
-        std::shared_ptr<TCursor> cursor = std::make_shared<TCursor>(TCursor(this));
-        AFL_VERIFY(Cursors.emplace(cursor).second);
-        return cursor;
-    }
-};
-
 class TDuplicateFilterConstructor: public NActors::TActor<TDuplicateFilterConstructor> {
 private:
     friend class TColumnFetchingContext;
