+++ ydb/core/tx/columnshard/engines/reader/simple_reader/duplicates/ut/ut_interval_counter.cpp
@@ -1,4 +1,4 @@
-#include <ydb/core/tx/columnshard/engines/reader/simple_reader/duplicates/manager.h>
+#include <ydb/core/tx/columnshard/engines/reader/simple_reader/duplicates/interval_counter.h>
 
 #include <library/cpp/testing/unittest/registar.h>
